bridge/requirements.txt
websockets>=13.0
pyserial>=3.5
nfcpy>=1.0.4
